+++ src/js/dropdown.js
@@ -9,6 +9,7 @@
         animationDuration: 300,
         animationType: 'none',
         hover: false,
+        autoClose: true,
       };
     }
 
@@ -21,7 +22,7 @@
     constructor(element, options, isLoadedWithData) {
       super();
 
-      Axentix.instances.push(this);
+      Axentix.instances.push({ type: 'Dropdown', instance: this });
 
       this.el = document.querySelector(element);
 
@@ -35,16 +36,13 @@
      */
     _setup() {
       Axentix.createEvent(this.el, 'dropdown.setup');
-      this.dropdownContent = document.querySelector('#' + this.el.id + ' .dropdown-content');
-      this.dropdownTrigger = document.querySelector('#' + this.el.id + ' .dropdown-trigger');
+
+      this.dropdownContent = this.el.querySelector('.dropdown-content');
+      this.dropdownTrigger = this.el.querySelector('.dropdown-trigger');
       this.isAnimated = false;
       this.isActive = this.el.classList.contains('active') ? true : false;
 
-      if (this.options.hover) {
-        this.el.classList.add('active-hover');
-      } else {
-        this._setupListeners();
-      }
+      this.options.hover ? this.el.classList.add('active-hover') : this._setupListeners();
 
       this._setupAnimation();
     }
@@ -124,6 +122,12 @@
       this.isActive ? this.close() : this.open();
     }
 
+    _autoClose() {
+      Axentix.getInstanceType('Dropdown').map((dropdown) => {
+        dropdown.el.id !== this.el.id ? dropdown.close() : '';
+      });
+    }
+
     /**
      * Open dropdown
      */
@@ -133,10 +137,13 @@
       }
       Axentix.createEvent(this.el, 'dropdown.open');
       this.dropdownContent.style.display = 'flex';
+
       setTimeout(() => {
         this.el.classList.add('active');
         this.isActive = true;
       }, 10);
+
+      this.options.autoClose ? this._autoClose() : '';
 
       if (this.options.animationType !== 'none') {
         this.isAnimated = true;