+++ src/js/fab.js
@@ -24,7 +24,7 @@
     constructor(element, options, isLoadedWithData) {
       super();
 
-      Axentix.instances.push(this);
+      Axentix.instances.push({ type: 'Fab', instance: this });
 
       this.el = document.querySelector(element);