+++ src/js/modal.js
@@ -21,7 +21,7 @@
     constructor(element, options, isLoadedWithData) {
       super();
 
-      Axentix.instances.push(this);
+      Axentix.instances.push({ type: 'Modal', instance: this });
 
       this.el = document.querySelector(element);