+++ src/js/tooltip.js
@@ -23,7 +23,7 @@
     constructor(element, options, isLoadedWithData) {
       super();
 
-      Axentix.instances.push(this);
+      Axentix.instances.push({ type: 'Tooltip', instance: this });
 
       this.el = document.querySelector(element);
       this.options = Axentix.getComponentOptions('Tooltip', options, this.el, isLoadedWithData);